Create Transform object to replace and eliminate KinematicGameObject
